Refactor code for type safety and improve resource loading in Bango Cat OBS plugin
This commit includes several type safety improvements by using static_cast for various variables across multiple files, ensuring better compatibility and reducing potential runtime errors. Additionally, the resource loading mechanism has been enhanced to allow for more flexible paths, improving the plugin's ability to locate necessary files regardless of the installation structure. The README.md has also been updated to reflect these changes.

+++ Pal.cpp
@@ -29,7 +29,7 @@ csmByte* Pal::LoadFileAsBytes(const string filePath, csmSizeInt* outSize)
     struct _stat64 statBuf;
     if (_wstat64(wstr, &statBuf) == 0)
     {
-        size = statBuf.st_size;
+        size = static_cast<int>(statBuf.st_size);
     } else {
         size = 0;
     }
